Check whether linebreak is needed during 'call'
Checking during teardown results in situations where the last test of a
file is the one that forces a linebreak for the next tests, but then the
file changes anyway, resulting in another linebreak.

Moving this check into 'call', right after checking for a new test file
solves this issue.