[travis] Install pypy
It is no longer installed on the new tusty images;
see travis-ci/travis-ci#8081

See https://travis-ci.org/michael-k/pytest-sugar/jobs/258546307 for a
build of the latest master failing with `InterpreterNotFound`.

For some reason python 3.6 is required as base.  See
https://travis-ci.org/michael-k/pytest-sugar/jobs/258546372 for a
build failing with python 3.5 as base and
https://travis-ci.org/michael-k/pytest-sugar/jobs/258556992 for a
build based on python 3.6 (otherwise unchanged).

Python 3.5 needs to be installed to prevent `InterpreterNotFound:
python3.5`.  This did not appear when I tested my commits with my
travis account, but when I opened the pull request.