Run tests on py27, py34, and pypy also against pytest 3.1
Minor versions of python3 claim to be downwards compatible.
pytest-sugar any pytest 3.1 are both tested against python 3.5 and
3.6, just not combined.  These two things in mind, testing the
combination on python 3.4 should be enough to cover python 3.5 and 3.6
too.

Skipping the combination of pytest-xdist<1.14 and pytest>=3.1 should
not affect many users, because users of pytest>=3.1 are unlikely to
use pytest-xdist<1.14.